Update dependencies and provide explanation for RELRO
Updated the versions of dependencies in go.sum and go.mod files to their latest versions to ensure optimal functionality and security of the application. This includes updates to github.com/fatih/color, github.com/mattn/go-isatty, and golang.org/x/sys

Modified the version variable in main.go leading to version upgrade from 1.0.1 to 1.1.0.

Clarified the meaning of the acronym "RELRO" (Relocation Read-Only) in main.go for better readability and understanding.

Corrected a grammatical error in the README.md file and updated the required Go version for building the project from 1.20 to 1.21.